% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_fp.R
\name{climate_fp}
\alias{climate_fp}
\title{download climate data}
\usage{
climate_fp(model, date, format = "https")
}
\arguments{
\item{model}{Gridded climate model names: "RTMA, URMA, CONSUS404, etc"}

\item{date}{Datetime}
}
\value{
downloads data, returns status
}
\description{
download climate data
}
\examples{
\dontrun{
r <- download_climate('RTMA',date,4,out_dir = ".")
}

}
