% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate}
\alias{download_climate}
\title{download climate data}
\usage{
download_climate(model, date, hours, out_dir = ".")
}
\arguments{
\item{model}{Gridded climate model names: "RTMA, URMA, CONSUS404, etc"}

\item{date}{Datetime}

\item{hours}{Number of hours to be downloaded following the start time}

\item{out_dir}{Where the files should be downloaded}
}
\value{
downloads data, returns status
}
\description{
download climate data
}
\examples{
\dontrun{
r <- download_climate('RTMA',date,4,out_dir = ".")
}

}
