% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog.R
\name{prog}
\alias{prog}
\title{download climate data}
\usage{
prog(p, recurs = FALSE)
}
\arguments{
\item{p}{number representing the percent progress}

\item{recurs}{determines whether or not the progress bar will print back onto the line}
}
\description{
download climate data
}
\examples{
\dontrun{
r <- download_climate('RTMA',date,4,out_dir = ".")
}

}
