% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_curvilinear_nc.R
\name{write_curvilinear_nc}
\alias{write_curvilinear_nc}
\title{Convert a netCDF to a stars object with a curvilinear grid}
\usage{
write_curvilinear_nc(
  out_fp,
  var,
  lat,
  long,
  time,
  var_name = "A1",
  var_units = "undefined",
  lat_name = "XLAT",
  long_name = "XLONG",
  time_name = "XTIME",
  time_units = "minutes since 1970-01-01 00:00:00"
)
}
\arguments{
\item{out_fp}{A netCDF filepath with a curvilinear grid.}

\item{var}{A string matching the variable inside of the NetCDF}

\item{lat}{A string matching the name of latitude in the NetCDF}

\item{long}{A string matching the name of longitude in the NetCDF}

\item{time}{A string matching the name of time in the NetCDF}

\item{var_name}{A string matching the variable inside of the NetCDF}

\item{var_units}{A string matching the variable inside of the NetCDF}

\item{lat_name}{A string matching the name of latitude in the NetCDF}

\item{long_name}{A string matching the name of longitude in the NetCDF}

\item{time_name}{A string matching the name of time in the NetCDF}

\item{time_units}{String defining time units}
}
\value{
A curvilinear \code{stars} object.
}
\description{
Convert a netCDF to a stars object with a curvilinear grid
}
\examples{
\dontrun{
r <- curvilinear_nc("hourlyScab/conus404/wrf2d_d01_1996-08-30_170000.nc","T2")
}
}
