% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvilinear_nc.R
\name{curvilinear_nc}
\alias{curvilinear_nc}
\title{Convert a netCDF to a stars object with a curvilinear grid}
\usage{
curvilinear_nc(
  fp,
  var_name,
  lat_name = "XLAT",
  long_name = "XLONG",
  time_name = "XTIME"
)
}
\arguments{
\item{fp}{A netCDF filepath with a curvilinear grid.}

\item{var_name}{A string matching the variable inside of the NetCDF}

\item{lat_name}{A string matching the name of latitude in the NetCDF}

\item{long_name}{A string matching the name of longitude in the NetCDF}

\item{time_name}{A string matching the name of time in the NetCDF}
}
\value{
A curvilinear \code{stars} object.
}
\description{
Convert a netCDF to a stars object with a curvilinear grid
}
\examples{
\dontrun{
r <- curvilinear_nc("hourlyScab/conus404/wrf2d_d01_1996-08-30_170000.nc","T2")
}
}
